# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SaltRiver
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /login:
    x-swagger-router-controller: user
    post:
      tags:
        - user
      description: 로그인하기
      operationId: login
      summary: 로그인하기
      parameters:
        - in: body
          name: admin
          description: The admin to create.
          schema:
            type: object
            properties:
              id:
                type: string
                description: 아이디
              password:
                type: string
                description: 비밀번호
      responses:
        "200":
          description: Login
          schema:
            $ref: "#/definitions/LoginResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user:
    x-swagger-router-controller: user
    get:
      tags:
        - user
      description: Find Firend
      operationId: findUser
      security:
        - jwt: []
      parameters:
        - name: search_text
          in: query
          description: search_text
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - user
      description: 회사를 처음으로 등록하는 사람은 자동으로 관리자 권한을 갖게 됩니다.
      operationId: addUser
      summary: 관리자를 등록할 때 사용하는 api
      parameters:
        - in: body
          name: admin
          description: 회원 가입시 입력하는 정보
          schema:
            type: object
            properties:
              info:
                type: object
                properties:
                  id:
                    type: string
                    description: 사용자가 입력한 id
                  pw:
                    type: string
                    description: 사용자가 입력한 비밀번호
                  name:
                    type: string
                    description: 사용자의 이름
                  email:
                    type: string
                    description: 사용자의 이메일
                  phone:
                    type: string
                    description: 사용자의 핸드폰 번호
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - user
      description: 회사를 처음으로 등록하는 사람은 자동으로 관리자 권한을 갖게 됩니다.
      operationId: patchUser
      security:
        - jwt: []
      summary: 유저 정보 수정
      parameters:
        - in: body
          name: admin
          description: 회원 가입시 입력하는 정보
          schema:
            type: object
            properties:
              info:
                type: object
                properties:
                  pw:
                    type: string
                    description: 사용자가 입력한 비밀번호
                  name:
                    type: string
                    description: 사용자의 이름
                  email:
                    type: string
                    description: 사용자의 이메일
                  phone:
                    type: string
                    description: 사용자의 핸드폰 번호
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /token:
    x-swagger-router-controller: user
    get:
      tags:
        - user
      description: 토큰에 대한 정보를 반환한다.
      operationId: checkToken
      security:
        - jwt: []
      summary: 토큰 유효성 검증
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /id:
    x-swagger-router-controller: user
    post:
      tags:
        - user
      description: 희망하는 아이디를 사용 가능한지 확인할 수 있는 함수
      operationId: checkID
      summary: 아이디 중복 체크
      parameters:
        - name: id
          in: query
          description: 사용하고자 하는 아이디
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
securityDefinitions:
  jwt:
    type: apiKey
    in: header
    name: jwt
# complex objects have schema definitions
definitions:
  LoginResponse:
    properties:
      data:
        type: object
        properties:
          token:
            type: string
            description: 성공 여부
          user_name:
            type: string
            description: 유저의 이름
          user_type:
            type: array
            description: 유저가 가진 회사 타입
            items :
              type : string
  SuccessResponse:
    properties:
      data:
        type: object
        properties:
          success:
            type: boolean
            description: 성공 여부
          message:
            type: string
            description: 메시지
  RelationResponse:
    properties:
      data:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string